+++ app/_hooks/use-render-context.ts
@@ -5,11 +5,13 @@ import { Layers } from "@/_lib/layers";
 export interface RenderContextType {
   erosions: number;
   layers: Layers;
+  onPageRenderSuccess: () => void;
 }
 
 export const RenderContext = createContext<RenderContextType>({
   erosions: 0,
   layers: {},
+  onPageRenderSuccess: () => {},
 });
 
 export default function useRenderContext() {