+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Export PDF button that saves the PDF with selected layers and stitched pages
 - Save visible layers by file per browser/device
 - Save position by file per browser/device
+- Status icons that indicate when a PDF is loading and for when line thickness is updated
 
 ### Changed